fixed missing cast

+++ lib/dslash_coarse.cu
@@ -1071,7 +1071,7 @@ namespace quda {
 		        dslash_policy == DslashCoarsePolicy::DSLASH_COARSE_GDR || 
             dslash_policy == DslashCoarsePolicy::DSLASH_COARSE_ZERO_COPY_PACK_GDR_RECV ||
 		        dslash_policy == DslashCoarsePolicy::DSLASH_COARSE_GDR_SEND_ZERO_COPY_READ) && !comm_gdr_enabled() ) {
-	      errorQuda("Cannot select a GDR policy %d unless QUDA_ENABLE_GDR is set", dslash_policy);
+	      errorQuda("Cannot select a GDR policy %d unless QUDA_ENABLE_GDR is set", static_cast<int>(dslash_policy));
 	    }
 
 	    enable_policy(dslash_policy);
